+++ apps/ledger-live-desktop/src/renderer/reducers/market.ts
@@ -1,48 +1,70 @@
 import { handleActions } from "redux-actions";
 import { Handlers } from "./types";
+import { MarketListRequestParams } from "@ledgerhq/live-common/market/types";
 
 export type MarketState = {
-  range?: string;
-  limit?: number;
-  ids?: string[];
-  starred?: string[];
-  orderBy?: string;
-  order?: string;
-  search?: string;
-  liveCompatible?: boolean;
-  page?: number;
-  counterCurrency?: string;
+  marketParams: {
+    range?: string;
+    limit?: number;
+    ids?: string[];
+    starred?: string[];
+    orderBy?: string;
+    order?: string;
+    search?: string;
+    liveCompatible?: boolean;
+    page?: number;
+    counterCurrency?: string;
+  };
+  starredMarketCoins: string[];
 };
 
 const initialState: MarketState = {
-  range: "24h",
-  limit: 50,
-  ids: [],
-  starred: [],
-  orderBy: "market_cap",
-  order: "desc",
-  search: "",
-  liveCompatible: false,
-  page: 1,
-  counterCurrency: "usd",
+  marketParams: {
+    range: "24h",
+    limit: 50,
+    ids: [],
+    starred: [],
+    orderBy: "market_cap",
+    order: "desc",
+    search: "",
+    liveCompatible: false,
+    page: 1,
+    counterCurrency: "usd",
+  },
+  starredMarketCoins: [],
 };
 
 type HandlersPayloads = {
-  MARKET_SET_VALUES: MarketState;
+  MARKET_SET_VALUES: MarketListRequestParams;
+  ADD_STARRED_MARKET_COINS: string;
+  REMOVE_STARRED_MARKET_COINS: string;
 };
+
 type MarketHandlers<PreciseKey = true> = Handlers<MarketState, HandlersPayloads, PreciseKey>;
 
 const handlers: MarketHandlers = {
-  MARKET_SET_VALUES: (state: MarketState, { payload }: { payload: MarketState }) => ({
+  MARKET_SET_VALUES: (state: MarketState, { payload }: { payload: MarketListRequestParams }) => ({
     ...state,
-    ...payload,
+    marketParams: {
+      ...state.marketParams,
+      ...payload,
+    },
+  }),
+  ADD_STARRED_MARKET_COINS: (state: MarketState, { payload }: { payload: string }) => ({
+    ...state,
+    starredMarketCoins: [...state.starredMarketCoins, payload],
+  }),
+  REMOVE_STARRED_MARKET_COINS: (state: MarketState, { payload }: { payload: string }) => ({
+    ...state,
+    starredMarketCoins: state.starredMarketCoins.filter(id => id !== payload),
   }),
 };
 
 // Selectors
 
-export const marketSelector = (state: { market: MarketState }) => state.market;
-
+export const marketParamsSelector = (state: { market: MarketState }) => state.market.marketParams;
+export const starredMarketCoinsSelector = (state: { market: MarketState }) =>
+  state.market.starredMarketCoins;
 // Exporting reducer
 
 export default handleActions<MarketState, HandlersPayloads[keyof HandlersPayloads]>(
