+++ apps/ledger-live-desktop/src/renderer/screens/market/components/SideDrawerFilter.tsx
@@ -1,14 +1,14 @@
 import React, { useCallback } from "react";
 import { TFunction } from "i18next";
 import Dropdown from "./DropDown";
-import { MarketState } from "~/renderer/reducers/market";
+import { MarketListRequestParams } from "@ledgerhq/live-common/market/types";
 
 export default function SideDrawerFilter({
   refresh,
   filters,
   t,
 }: {
-  refresh: (params: MarketState) => void;
+  refresh: (params: MarketListRequestParams) => void;
   filters: Record<
     "starred" | "liveCompatible",
     { value: boolean; toggle: () => void; disabled?: boolean }
